adapt navigation
